pyproject.toml
[tool.poetry]
name = "python-django-team27"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "market"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.3"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.6"
ipython = "^8.14.0"
django-extensions = "^3.2.3"
dj-database-url = "^2.0.0"
django-jinja = "^2.10.2"
pillow = "^10.0.0"
redis-om = "^0.2.1"
config = "^0.5.1"
django-constance = {extras = ["database"], version = "^3.1.0"}
gloriousprotonmanager = "^0.80.1"
django-celery-beat = "^2.5.0"
django-celery-results = "^2.5.1"
pyyaml = "^6.0.1"
requests = "^2.31.0"
celery = "^5.3.4"
django-filter = "^23.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.7.0"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line_length = 119
